jobs/mktdata-archiver/pyproject.toml
[tool.poetry]
name = "mktdata-archiver"
version = "1.0"
description = "Market Data Archiver Service"
authors = ["Muwazana S.A. <exchanges@muwazana.com>"]

[[tool.poetry.source]]
name = "mwz"
url = "https://nexus.tools.muwazana.com/repository/muwazana-pypi/simple"

[tool.poetry.dependencies]
python = "^3.7"
# paf = "^1.5.0"
paf = { path = "../../libs/paf" }
# mwzpyutil = "^1.5.5"
mwzpyutil = { path = "../../libs/mwzpyutil" }
pyarrow = "^0.16.0"
google-cloud-storage = "^1.26.0"

[tool.poetry.dev-dependencies]
black = { version = "*", allow-prereleases = true }
ipython = "^7.13.0"
ipdb = "^0.13.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
